TEST) CI/CD TEST

See merge request s09-webmobile1-sub2/S09P12A703!30

+++ gaebap/src/main/java/site/doggyyummy/gaebap/domain/meeting/dto/ModifyMeetingResponseDTO.java
@@ -18,7 +18,7 @@ public class ModifyMeetingResponseDTO {
     private String password;
 
     @JsonProperty("max_participant")
-    private Long maxParticipant;
+    private int maxParticipant;
 
     private String title;
